[FLINK-19492][core] Consolidate Source Events between Source API and Split Reader API
