+++ flink-core/src/main/java/org/apache/flink/api/connector/source/event/RequestSplitEvent.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.connector.base.source.event;
+package org.apache.flink.api.connector.source.event;
 
 import org.apache.flink.api.connector.source.SourceEvent;
 
@@ -26,6 +26,9 @@ import java.util.Objects;
 
 /**
  * An event to request splits, sent typically from the Source Reader to the Source Enumerator.
+ *
+ * <p>This event optionally carries the hostname of the location where the reader runs, to support
+ * locality-aware work assignment.
  */
 public final class RequestSplitEvent implements SourceEvent {
 
@@ -34,6 +37,16 @@ public final class RequestSplitEvent implements SourceEvent {
 	@Nullable
 	private final String hostName;
 
+	/**
+	 * Creates a new {@code RequestSplitEvent} with no host information.
+	 */
+	public RequestSplitEvent() {
+		this(null);
+	}
+
+	/**
+	 * Creates a new {@code RequestSplitEvent} with a hostname.
+	 */
 	public RequestSplitEvent(@Nullable String hostName) {
 		this.hostName = hostName;
 	}
