+++ flink-core/src/main/java/org/apache/flink/api/connector/source/lib/util/IteratorSourceEnumerator.java
@@ -21,6 +21,8 @@ package org.apache.flink.api.connector.source.lib.util;
 import org.apache.flink.api.connector.source.SourceEvent;
 import org.apache.flink.api.connector.source.SplitEnumerator;
 import org.apache.flink.api.connector.source.SplitEnumeratorContext;
+import org.apache.flink.api.connector.source.event.NoMoreSplitsEvent;
+import org.apache.flink.api.connector.source.event.RequestSplitEvent;
 import org.apache.flink.util.FlinkRuntimeException;
 
 import java.util.ArrayDeque;
@@ -56,7 +58,7 @@ public class IteratorSourceEnumerator<SplitT extends IteratorSourceSplit<?, ?>> 
 
 	@Override
 	public void handleSourceEvent(int subtaskId, SourceEvent sourceEvent) {
-		if (!(sourceEvent instanceof SplitRequestEvent)) {
+		if (!(sourceEvent instanceof RequestSplitEvent)) {
 			throw new FlinkRuntimeException("Unrecognized event: " + sourceEvent);
 		}
 
@@ -64,7 +66,7 @@ public class IteratorSourceEnumerator<SplitT extends IteratorSourceSplit<?, ?>> 
 		if (nextSplit != null) {
 			context.assignSplit(nextSplit, subtaskId);
 		} else {
-			context.sendEventToSourceReader(subtaskId, new NoSplitAvailableEvent());
+			context.sendEventToSourceReader(subtaskId, new NoMoreSplitsEvent());
 		}
 	}
